requirements.txt
yfinance>=0.2.36
pandas>=2.2.0
PyGithub>=2.1.1